Replaced GeneticAlgorithmTemplate and its implementations with new EvolutionEngine concept
