+++ src/main/scala/genetic/engines/bestTracking/package.scala
@@ -1,0 +1,7 @@
+package genetic.engines
+
+import genetic.Population
+
+package object bestTracking {
+  type PopulationWithBest[G] = (Population[G], (G, Double))
+}
