--- src/main/scala/genetic/engines/streaming/package.scala
@@ -1,12 +1,0 @@
-package genetic.engines
-
-import akka.stream.ActorMaterializer
-import genetic.engines.sync.SynchronousGA
-
-import scala.concurrent.ExecutionContext
-
-package object streaming {
-  implicit class StreamingSupport(private val syncGA: SynchronousGA) extends AnyVal {
-    def stream(implicit mat: ActorMaterializer, ex: ExecutionContext): StreamingGA = new StreamingGA(syncGA)
-  }
-}
